updated for pyqt5 and python 3; in progress
