+++ GraphWindow.py
@@ -2,15 +2,17 @@
 Window for holding Graphs
 '''
 import sys
-from PyQt4 import QtGui
 import GUIConfig
+
+from PyQt5 import QtCore, QtWidgets
+
 from GraphWidgetPyQtGraph import Graph_PyQtGraph as Graph
 from HistWidgetPyQtGraph import Hist_PyQtGraph as Hist
 from ScrollingGraphWidgetPyQtGraph import ScrollingGraph_PyQtGraph as ScrollingGraph
 from ImageWidget import imageWidget as ImageGraph
 from GridGraphWindow import GridGraphWindow
 
-class GraphWindow(QtGui.QTabWidget):
+class GraphWindow(QtWidgets.QTabWidget):
     def __init__(self, reactor, cxn = None, parent=None):
         super(GraphWindow, self).__init__(parent)
         self.cxn = cxn
@@ -62,9 +64,12 @@ class GraphWindow(QtGui.QTabWidget):
         self.reactor.stop()
 
 if __name__ == '__main__':
-    app = QtGui.QApplication(sys.argv)
-    import qt4reactor
-    qt4reactor.install()
+    app = QtWidgets.QApplication(sys.argv)
+    # import qt5reactor
+    # qt5reactor.install()
+    # from twisted.internet import reactor
+    # from twisted.application import reactors
+    # reactors.installReactor('qt4')
     from twisted.internet import reactor
     main = GraphWindow(reactor)
     main.show()
