+++ GridGraphWindow.py
@@ -2,7 +2,7 @@
 Window containing a grid of graphs
 '''
 import sys
-from PyQt4 import QtGui
+from PyQt5 import QtWidgets
 from GraphWidgetPyQtGraph import Graph_PyQtGraph as Graph
 from ScrollingGraphWidgetPyQtGraph import ScrollingGraph_PyQtGraph as ScrollingGraph
 import GUIConfig
@@ -10,7 +10,7 @@ from twisted.internet.defer import Deferred, inlineCallbacks, returnValue
 from twisted.internet.task import LoopingCall
 from twisted.internet.threads import blockingCallFromThread
 
-class GridGraphWindow(QtGui.QWidget):
+class GridGraphWindow(QtWidgets.QWidget):
     def __init__(self, g_list, row_list, column_list, reactor, parent=None):
         super(GridGraphWindow, self).__init__(parent)
         self.reactor = reactor        
@@ -19,7 +19,7 @@ class GridGraphWindow(QtGui.QWidget):
 
     def initUI(self, g_list, row_list, column_list):
         reactor = self.reactor
-        layout = QtGui.QGridLayout()
+        layout = QtWidgets.QGridLayout()
         for k in range(len(g_list)):
             layout.addWidget(g_list[k], row_list[k], column_list[k])
         self.setLayout(layout)
