+++ HistWidgetPyQtGraph.py
@@ -1,11 +1,11 @@
 import sys
-from PyQt4 import QtGui, QtCore
+from PyQt5 import QtWidgets, QtCore, QtGui
 import pyqtgraph as pg
 from TraceListWidget import TraceList
 from twisted.internet.defer import inlineCallbacks, returnValue
 from twisted.internet.task import LoopingCall
 import itertools
-import Queue
+import queue
 
 
 class artistParameters():
@@ -19,7 +19,7 @@ class artistParameters():
                               # update count
 
                               
-class Hist_PyQtGraph(QtGui.QWidget):
+class Hist_PyQtGraph(QtWidgets.QWidget):
     def __init__(self, config, reactor, cxn=None, parent=None):
         super(Hist_PyQtGraph, self).__init__(parent)
         self.cxn = cxn
@@ -59,13 +59,13 @@ class Hist_PyQtGraph(QtGui.QWidget):
             init_value = yield self.get_init_vline()
             self.inf.setValue(init_value)
             self.inf.setPen(width=5.0)
-        self.coords = QtGui.QLabel('')
-        self.title = QtGui.QLabel(self.name)
-        frame = QtGui.QFrame()
-        splitter = QtGui.QSplitter()
+        self.coords = QtWidgets.QLabel('')
+        self.title = QtWidgets.QLabel(self.name)
+        frame = QtWidgets.QFrame()
+        splitter = QtWidgets.QSplitter()
         splitter.addWidget(self.tracelist)
-        hbox = QtGui.QHBoxLayout()
-        vbox = QtGui.QVBoxLayout()
+        hbox = QtWidgets.QHBoxLayout()
+        vbox = QtWidgets.QVBoxLayout()
         vbox.addWidget(self.title)
         vbox.addWidget(self.pw)
         vbox.addWidget(self.coords)
@@ -133,7 +133,7 @@ class Hist_PyQtGraph(QtGui.QWidget):
             except KeyError:
                 pass
         except:
-            print "remove failed"
+            print("remove failed")
 
     def display(self, ident, shown):
         try:
@@ -197,7 +197,7 @@ class Hist_PyQtGraph(QtGui.QWidget):
     def get_init_vline(self):
         init_vline = yield self.pv.get_parameter(self.vline_param[0],
                                                  self.vline_param[1])
-        print init_vline
+        print(init_vline)
         returnValue(init_vline)
 
     @inlineCallbacks
@@ -209,9 +209,9 @@ class Hist_PyQtGraph(QtGui.QWidget):
 
 
 if __name__ == '__main__':
-    app = QtGui.QApplication(sys.argv)
-    import qt4reactor
-    qt4reactor.install()
+    app = QtWidgets.QApplication(sys.argv)
+    import qt5reactor
+    qt5reactor.install()
     from twisted.internet import reactor
     main = Hist_PyQtGraph('example', reactor)
     main.show()
