+++ PredictSpectrumWidget.py
@@ -1,6 +1,6 @@
-# import qt4reactor
-# qt4reactor.install()
-from PyQt4 import QtGui, QtCore
+import qt5reactor
+qt5reactor.install()
+from PyQt5 import QtWidgets, QtCore
 from twisted.internet.defer import inlineCallbacks, returnValue, DeferredLock, Deferred
 from fractions import Fraction
 # from labrad import units as U
@@ -16,7 +16,7 @@ class ParamInfo():
     def __init__(self, value):
         self.value = value
 
-class PredictSpectrum(QtGui.QWidget):
+class PredictSpectrum(QtWidgets.QWidget):
 
     def __init__(self, parent):
         super(PredictSpectrum, self).__init__()
@@ -29,31 +29,31 @@ class PredictSpectrum(QtGui.QWidget):
 
     def initUI(self):
         self.setWindowTitle(self.ident)
-        mainLayout = QtGui.QVBoxLayout()
-        buttons = QtGui.QHBoxLayout()
+        mainLayout = QtWidgets.QVBoxLayout()
+        buttons = QtWidgets.QHBoxLayout()
 
-        self.parameterTable = QtGui.QTableWidget()
+        self.parameterTable = QtWidgets.QTableWidget()
         self.parameterTable.setColumnCount(2)
 
-        self.plotButton = QtGui.QPushButton('Plot', self)
+        self.plotButton = QtWidgets.QPushButton('Plot', self)
 
         mainLayout.addWidget(self.parameterTable)
         mainLayout.addLayout(buttons)
         buttons.addWidget(self.plotButton)
 
-        self.OPpos = QtGui.QCheckBox("Positive Manifold")
+        self.OPpos = QtWidgets.QCheckBox("Positive Manifold")
         self.OPpos.setChecked(True)
         mainLayout.addWidget(self.OPpos)
-        self.OPneg = QtGui.QCheckBox("Negative Manifold")
+        self.OPneg = QtWidgets.QCheckBox("Negative Manifold")
         self.OPneg.setChecked(True)
         mainLayout.addWidget(self.OPneg)
-        self.deltam0 = QtGui.QCheckBox("Delta m=0")
+        self.deltam0 = QtWidgets.QCheckBox("Delta m=0")
         self.deltam0.setChecked(True)
         mainLayout.addWidget(self.deltam0)
-        self.deltam1 = QtGui.QCheckBox("Delta m=1")
+        self.deltam1 = QtWidgets.QCheckBox("Delta m=1")
         self.deltam1.setChecked(True)
         mainLayout.addWidget(self.deltam1)
-        self.deltam2 = QtGui.QCheckBox("Delta m=2")
+        self.deltam2 = QtWidgets.QCheckBox("Delta m=2")
         self.deltam2.setChecked(True)
         mainLayout.addWidget(self.deltam2)
 
@@ -66,8 +66,7 @@ class PredictSpectrum(QtGui.QWidget):
     def setupParameterTable(self):
 
         self.parameterTable.clear()
-        
-        headerLabels = QtCore.QStringList(['Parameter', 'Value'])
+        headerLabels = ['Parameter', 'Value']
         self.parameterTable.setHorizontalHeaderLabels(headerLabels)
         self.parameterTable.horizontalHeader().setStretchLastSection(True)
 
@@ -75,8 +74,8 @@ class PredictSpectrum(QtGui.QWidget):
         self.parameterTable.setRowCount(len(params))
         for i,p in enumerate(params):
 
-            label = QtGui.QLabel(p)
-            value = QtGui.QDoubleSpinBox()
+            label = QtWidgets.QLabel(p)
+            value = QtWidgets.QDoubleSpinBox()
 
             self.value_dict[p] = ParamInfo(value)
 
@@ -102,7 +101,7 @@ class PredictSpectrum(QtGui.QWidget):
 
             all_carriers = self.Ca_data.get_transition_energies(b_field*1e-4,line_center) #to Tesla and MHz
 
-            print all_carriers
+            print(all_carriers)
 
             #choose which carriers to include
             included_lines = []
@@ -207,7 +206,7 @@ class EnergyLevel(object):
         J = total_angular_momentum_j
         lande_factor =  self.lande_factor(S, L, J)
         #sublevels are found, 2* self.J is always an integer, so can use numerator
-        self.sublevels_m =  [-J + i for i in xrange( 1 + (2 * J).numerator)]
+        self.sublevels_m =  [-J + i for i in range( 1 + (2 * J).numerator)]
         self.energy_scale = (lande_factor * 9.274e-24 / 6.626e-34) #1.4 MHz / gauss
     
     def lande_factor(self, S, L ,J):
