+++ ScrollingGraphWidgetPyQtGraph.py
@@ -1,5 +1,5 @@
 from GraphWidgetPyQtGraph import Graph_PyQtGraph as Graph
-from PyQt4 import QtGui, QtCore
+from PyQt5 import QtGui, QtCore, QtWidgets
 
 class ScrollingGraph_PyQtGraph(Graph):
     def __init__(self, name, reactor, parent = None, ylim=[0,1], cxn=None):
@@ -20,9 +20,9 @@ class ScrollingGraph_PyQtGraph(Graph):
                 
             
         try:
-    	    mousepressed =  QtGui.qApp.mouseButtons()
-    	    if (mousepressed == QtCore.Qt.LeftButton) or (mousepressed == QtCore.Qt.RightButton):
-                return 
+            mousepressed = QtWidgets.QApplication.instance().MouseButton
+            if mousepressed == (QtCore.Qt.LeftButton or QtCore.Qt.RightButton):
+                return
                 # see if we need to redraw
             xmin_cur, xmax_cur = self.current_limits
             x_cur = x[-1] # current largest x value
