+++ TraceListWidget.py
@@ -1,12 +1,11 @@
-from PyQt4 import QtGui
-from PyQt4 import QtCore
+from PyQt5 import QtGui, QtWidgets, QtCore
 from ParameterListWidget import ParameterList
 from DataVaultListWidget import DataVaultList
 from FitWindowWidget import FitWindow
 from PredictSpectrumWidget import PredictSpectrum
 from GUIConfig import traceListConfig
 
-class TraceList(QtGui.QListWidget):
+class TraceList(QtWidgets.QListWidget):
     def __init__(self, parent):
         super(TraceList, self).__init__()
         self.parent = parent
@@ -23,14 +22,14 @@ class TraceList(QtGui.QListWidget):
 
     def initUI(self):
         self.trace_dict = {}
-        item = QtGui.QListWidgetItem('Traces')
+        item = QtWidgets.QListWidgetItem('Traces')
         item.setCheckState(QtCore.Qt.Checked)
         self.setContextMenuPolicy(QtCore.Qt.CustomContextMenu)
         self.customContextMenuRequested.connect(self.popupMenu)
 
 
     def addTrace(self, ident, color):
-        item = QtGui.QListWidgetItem(ident)
+        item = QtWidgets.QListWidgetItem(ident)
 
         if self.use_trace_color:
             foreground_color = self.parent.getItemColor(color)
@@ -54,7 +53,7 @@ class TraceList(QtGui.QListWidget):
         item.setForeground(self.parent.getItemColor(new_color))
 
     def popupMenu(self, pos):
-        menu = QtGui.QMenu()
+        menu = QtWidgets.QMenu()
         item = self.itemAt(pos)
         if (item == None): 
             dataaddAction = menu.addAction('Add Data Set')
