+++ analysis/fit_sinusoid.py
