+++ analysis/fit_sinusoid2.py
