+++ analysis/fitting.py
@@ -2,19 +2,19 @@
 
 import numpy as np
 from scipy import optimize
-from fit_lorentzian import Lorentzian
-from fit_gaussian import Gaussian
-from fit_linear import Linear
-from fit_rabi import Rabi
-from fit_bessel import Bessel
-from fit_rotrabi import RotRabi
-from fit_rotramsey import RotRamsey
-from fit_sinusoid import Sinusoid
-from fit_sinusoid2 import Sinusoid2
-from fit_expdecay import ExponentialDecay
-from fit_gaussdecay import GaussianDecay
-from fit_ramsey import RamseyDecay
-from fit_ramseybfield import RamseyBfield
+# from fit_lorentzian import Lorentzian
+# from fit_gaussian import Gaussian
+# from fit_linear import Linear
+# from fit_rabi import Rabi
+# from fit_bessel import Bessel
+# from fit_rotrabi import RotRabi
+# from fit_rotramsey import RotRamsey
+# from fit_sinusoid import Sinusoid
+# from fit_sinusoid2 import Sinusoid2
+# from fit_expdecay import ExponentialDecay
+# from fit_gaussdecay import GaussianDecay
+# from fit_ramsey import RamseyDecay
+# from fit_ramseybfield import RamseyBfield
 
 class FitWrapper():
