+++ depricated/ScrollingGraphWidget.py
