+++ rsg.py
@@ -1,19 +1,18 @@
 '''
 The Real Simple Grapher
 '''
-
+import qt5reactor
 from GraphWindow import GraphWindow
 from Dataset import Dataset
-from PyQt4 import QtGui
-a = QtGui.QApplication( [])
-import qt4reactor
-qt4reactor.install()
+from PyQt5 import QtWidgets
+a = QtWidgets.QApplication([])
+# import qt5reactor
+# qt5reactor.install()
 #import server libraries
 from twisted.internet.defer import returnValue, DeferredLock, Deferred, inlineCallbacks
 from twisted.internet.threads import deferToThread
 from twisted.internet import reactor
 from labrad.server import LabradServer, setting
-
 
 """
 ### BEGIN NODE INFO
