+++ tests/cursor_test.py
@@ -8,6 +8,6 @@ vb = w.addViewBox()
 img = pg.ImageItem(np.random.normal(size=(100,100)))
 vb.addItem(img)
 def mouseMoved(pos):
-    print "Image position:", img.mapFromScene(pos)
+    print("Image position:", img.mapFromScene(pos))
     
 w.scene().sigMouseMoved.connect(mouseMoved)
